Implement feedback from 2nd internal Meeting
Institute section:

ADD Prof. Ossowski as a contact person for the institute of medical genetics and applied genomics card
REMOVE image and name of Prof. Autenrieth from the institute of medical genetics and applied genomics card
ADD image and name of Prof. Frick to the institute of medical genetics and applied genomics card
ADD Janina Geißert as a contact person to the institute of medical microbiology and hygiene card
CC-Network section
ADD Central Coordinating Unit as a list entry with corresponding link and logo